Reorganized project into models, endpoints and daos. Added required api dependency for log4j. Added template file and background file resources

+++ src/main/java/com/github/eoinf/jiggen/PuzzleRepository.java
@@ -3,8 +3,9 @@ package com.github.eoinf.jiggen;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called puzzleRepository
 // CRUD refers Create, Read, Update, Delete
 
+import com.github.eoinf.jiggen.data.CachedPuzzle;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PuzzleRepository extends CrudRepository<CachedPuzzle, Long> {
+public interface PuzzleRepository extends CrudRepository<CachedPuzzle, Integer> {
 
 }
